+++ src/utils/lib/api.ts
@@ -1,3 +1,4 @@
+import axios from "axios";
 import useSWR from "swr";
 
 import { BookcaseInfo } from "../../pages/Bookcase";
@@ -7,11 +8,6 @@ import { client, KAKAO } from ".";
 export const searchBook = (params: KAKAOParams) => {
   return KAKAO.get("/v3/search/book", { params });
 };
-
-// headers에 들어갈 내용의 예시
-// "Content-Type": "application/json",
-// "Content-Type": "multipart/form-data"
-// "Authorization": "토큰"
 
 export const getData = (key: string, token?: string) => {
   return client(token).get(key);
@@ -61,3 +57,28 @@ export function useGetBookInfo(key: string) {
     isError: error,
   };
 }
+
+export const checkIsBookExist = async (isbn: string) => {
+  const _token = localStorage.getItem("booktez-token");
+  const userToken = _token ? _token : "";
+
+  try {
+    const { data } = await client(userToken).get(`/book/exist/${isbn}`);
+
+    if (data.success) {
+      return { isError: false, isExist: data.data.isExist };
+    } else {
+      // 통신에는 성공했으나 에러가 난 경우
+      console.log("[ERROR RETURNED]", data);
+
+      return { isError: true, isExist: false };
+    }
+  } catch (err) {
+    // 통신에 실패한 경우
+    if (axios.isAxiosError(err)) {
+      console.log("[ERROR CATCHED] statusCode: ", err.response?.status, err.message);
+    }
+
+    return { isError: true, isExist: false };
+  }
+};
